% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_winsorize.R
\name{add_winsorize}
\alias{add_winsorize}
\title{Add Winsorized (Outlier-Clipped) Factors}
\usage{
add_winsorize(
  data,
  cols = NULL,
  by = "date",
  probs = c(0.01, 0.99),
  append = TRUE,
  new_col = "win"
)
}
\arguments{
\item{data}{data.frame or tibble in standard long format}

\item{cols}{vector of column names to winsorize}

\item{by}{cross-sectional grouping column (default = "date")}

\item{probs}{lower and upper quantile limits (default = c(0.01, 0.99))}

\item{append}{if TRUE, append results to input data}

\item{new_col}{prefix for new columns}
}
\value{
A tibble or data frame with original columns and winsorized factors:
\describe{
\item{date}{Trading date}
\item{code}{Stock code}
\item{...}{All original input columns}
\item{\code{win_xxx}}{Winsorized factor (outliers clipped to quantile limits)}
}
}
\description{
Add cross-sectionally winsorized columns to limit the impact of extreme outliers.
}
