% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sample_stock_data}
\alias{sample_stock_data}
\title{Sample Stock Data}
\format{
A data frame with 5 columns:
\describe{
\item{code}{Stock ticker symbol}
\item{date}{Trading date}
\item{close}{Closing price}
\item{volume}{Trading volume}
\item{return}{Daily simple return}
}
}
\source{
Yahoo Finance using quantmod.
}
\usage{
sample_stock_data
}
\description{
A dataset containing daily price and return information for testing.
}
\keyword{datasets}
