% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_winsorize.R
\name{add_winsorize}
\alias{add_winsorize}
\title{Add Winsorized (Outlier-Clipped) Factors}
\usage{
add_winsorize(
  data,
  cols = NULL,
  by = "date",
  probs = c(0.01, 0.99),
  append = TRUE,
  new_col = "win"
)
}
\arguments{
\item{data}{data.frame or tibble in standard long format}

\item{cols}{vector of column names to winsorize}

\item{by}{cross-sectional grouping column (default = "date")}

\item{probs}{lower and upper quantile limits (default = c(0.01, 0.99))}

\item{append}{if TRUE, append results to input data}

\item{new_col}{prefix for new columns}
}
\value{
A tibble/data.frame with original columns plus new winsorized columns:
\itemize{
\item \strong{date}: trading date
\item \strong{code}: stock code
\item \strong{...}: original input columns
\item \strong{win_xxx}: winsorized version of input column xxx
}
}
\description{
Add cross-sectionally winsorized columns to limit the impact of extreme outliers.
}
